o2c-app/backend/nl-to-sql/requirements.txt
fastapi
uvicorn
python-dotenv
langchain>=0.3
langchain-groq
langchain-community
pymysql
sqlalchemy
